refactor: 모달 리팩터링에 따른 컴포넌트 적용 및 경로 수정
